# public/api-doc.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/mine-92b/catway_api/1.0.0
info:
  version: "1.0.0"
  title: Catway-api
  description: The API for the catway management and reservation in Russell harbor.
paths:
  /users:  
    post:
      tags:
        - Users
      description: Create a new user in the database
      operationId: addUser
      security:
        - cookieAuth: []
      parameters:
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lastName
                - firstName
                - email
                - password
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                  description: Must be unique, two users can't have the same email
                password:
                  type: string
              example:
                firstName: "Jean"
                lastName: "Dupont"
                email: "jean-dupont@gmail.com"
                password: "strong-password1234"
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "L'utilisateur Jean Dupont a bien été ajouté"
        '403':
          description: Email already used by a registered user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Un utilisateur utilise déjà cette adresse mail"
        '400':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "L'utilisateur Jean Dupont n'a pas pu être ajouté. Assurez-vous d'avoir rempli tous les champs."
...
  /catways:
    post:
      tags:
        - Catways
      description: Add new catway to the database
      operationId: addCatway
      security:
        - cookieAuth: []
      parameters:
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - catwayNumber
                - catwayState
                - type
              properties:
                catwayNumber:
                  description: Two catways can't have the same catway number
                  type: integer
                  format: int32
                  minimum: 1
                catwayState:
                  type: string
                type:
                  type: string
                  enum:
                    - short
                    - long
                  default: short
              example: 
                catwayNumber: 1
                catwayState: "bon état"
                type: "short"
      responses:
        '200':
          description: Catway successfully created into the database
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Le catway a bien été enregistré"
        '403':
          description: A registered catway already has this catway number
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Un catway possède déjà le même numéro"
        '400':
          description: An unexpected error occurs, catway is not created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Le catway n'a pas pu être ajouté. Assurez-vous d'avoir rempli tous les champs."
...
  '/catways/{id}/reservations/{idReservation}':
    get:
      tags:
        - Reservations
      description: Returns a specific reservation by ID from the database
      operationId: getReservation
      security:
        - cookieAuth: []
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: idReservation
          required: true
          in: path
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      responses:
        '200':
          description: Reservation successfully retrieved in the database
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  catwayNumber:
                    type: integer
                    format: int32
                    minimum: 1
                  type:
                    type: string
                    enum:
                      - short
                      - long
                  catwayState:
                    type: string
                example:
                  _id: 686101205b42028090c2a1a6
                  catwayNumber: 1
                  type: "short"
                  catwayState: "bon état"
        '404':
         description: Not found - data not found because of a wrong ID
         content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
            examples:
              reservationNotFound:
                summary: No reservation found with the provided ID
                value:
                  message: "L'identifiant de la réservation ne correspond à aucune réservation pour ce catway."
              catwayNotFound:
                summary: No catway found with the provided ID
                value:
                  message: "L'identifiant du catway ne correspond à aucun catway."
        '501':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
    patch:
      description: Updates a specific reservation data by the ID
      tags:
        - Reservations
      operationId: updateReservation
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: idReservation
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - catwayNumber
                - clientName
                - boatName
                - checkIn
                - checkOut
              properties:
                catwayNumber:
                  type: integer
                  format: int32
                  minimum: 1
                clientName:
                  type: string
                boatName:
                  type: string
                checkIn:
                  type: string
                  format: date-time
                checkOut:
                  type: string
                  format: date-time
              example:
                catwayNumber: 1
                clientName: "John Doe"
                boatName: "The Boat"
                checkIn: "2025-07-15T14:00:00.000Z"
                checkOut: "2025-08-14T13:00:00.000Z"
      responses:
        '204':
          description: Reservation successfully updated in the database
        '400':
          description: A required request body field is missing
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Tous les champs sont requis."
        '404':
          description: Element not found with provided data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                reservationNotFound:
                  summary: Reservation not found with provided ID
                  value:
                    message: "Tous les champs sont requis."
                catwayNotFoundWithID:
                  summary: Catway not found with provided ID
                  value:
                    message: "L'identifiant de catway renseigné dans l'URL ne correspond à aucun catway."
                catwayNotFoundWithNumber:
                  summary: Catway not found with provided number
                  value:
                    message: "Attention ce numéro de catway ne correspond à aucun catway."
        '403':
          description: Forbidden – reservation rejected for one of several reasons
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type : string
              examples:
                checkInAlreadyReserved:
                  summary: Check-in is during a period already reserved
                  value:
                    message: "L'arrivée a lieu sur une réservation déjà enregistrée."
                checkOutAlreadyReserved:
                  summary: Check-out is during a period already reserved
                  value:
                    message: "Le départ a lieu sur une réservation déjà enregistrée."
                periodAlreadyReserved:
                  summary: There is a period already reserved between check-in and check-out
                  value:
                    message: "Une réservation est déjà enregistrée au cours de la période demandée."
                invalidCheckIn:
                  summary: Check-in can't be before now
                  value:
                    message: "Attention la date et l'heure d'arrivée ne peuvent pas être antérieures à maintenant."
                invalidCheckOut:
                  summary: Check-out can't be before check-in
                  value:
                    message: "La date et l'heure de départ ne peuvent pas être antérieures à l'heure et la date d'arrivée."
        '501':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
    delete:
      tags:
        - Reservations
      description: Deletes a specific reservation into the database
      operationId: deleteReservation
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: idReservation
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      responses:
        '204':
          description: Catway successfully deleted
        '404':
          description: Element not found with provided data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                reservationNotFound:
                  summary: Reservation not found with provided ID
                  value:
                    message: "Cet identifiant ne correspond à aucune réservation faite sur ce catway."
                catwayNotFoundWithID:
                  summary: Catway not found with provided ID
                  value:
                    message: "L'identifiant de catway renseigné dans l'URL ne correspond à aucun catway."
        '403':
          description: The ID must contain 24 characters including letters and numbers
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "L'identifiant doit faire 24 caractères et doit contenir des chiffres et des lettres."
components:
  schemas:
    ApiResponse501:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Nous ne parvenons pas à nous connecter à notre base de données. Veuillez réessayer ultérieurement."
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token